requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
comtypes==1.1.10
Cython==0.29.23
Flask==2.0.1
Flask-Cors==3.0.10
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
keyboard==0.13.5
MarkupSafe==2.0.1
MouseInfo==0.1.3
numpy==1.21.0
opencv-python==4.5.2.54
Pillow==8.3.0
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
pytesseract==0.3.8
PyTweening==1.0.3
pywin32==301
pywinauto==0.6.8
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
tesseract==0.1.3
urllib3==1.26.6
Werkzeug==2.0.1
